Improvements to template anc vfc_ci_serve wrapper
The vfc_ci_serve wrapper now accepts arguments: in particular, it is now
possible to specify a git repository (either locally or by URL) hosted
on Github or Gitlab. When specified, the template will include links to
the different commits (directly on the plots) and to the repo's main
page.