Server code refactor
The runs comparison has been moved to a dedicated file, and put in a
class. This approach will be useful to make the runs and variables
comparisons live together (and potentially interact).

The workflow has also been sligthly modified to try to put the commit's
hash in the artifact name.