Inspect runs controller, and views master
Most of the "controller" (ie. widgets, callbacks, ...) for the run
inspection view is complete. The current run still needs to be updated
when selecting a glypth in the runs comparison. This should be easy to
implement by having a server-side callback on a glyph tap, an by using
the new ViewsMaster class. It is a container for view classes (we have 2
so far but this should be easy to add new ones in the future), and
allows communications between them.