Server side callback on glyph selection, unified runs names
The server side callback is now triggered to update "Inspect runs" selection,
when clicking a glyph in "Compare runs". Since it was easier
to pass the run's name inside of its timestamp, this was the occasion
to generate run names (x-axis ticks) in ViewsMaster : this way, we can
enforce their unicity and homogeneity across the entire report.