Sigma and s plots for run inspection
Added boxplots of aggregated sigma or s, where each box corresponds
to either a backend or a variable depending on the selected mode (Select
variable or backend). The ColumnDataSources used for those plots also
contain the aggragated average, which can probably be represented as a
barplot. This will probably be added in the next commit.