Replace bar plots with dot plots
Bar plots in both runs comparison & inspection have been replaced with
dot plots.
Since dot plots in runs comparison are in chronological order, there's a
line drawn between dots to better appreciate the evolution of metrics.
This is not the case in the runs inspection where this would not make
much sense.