New vfc_ci serve option to display logo in report
This option specifies an image URL that will be integrated in the
template and displayed in the report header.
This was the occasion to improve the way the args are parsed in the
server's main.py, so it should be easy to add new options that modify
the template or require some processing in main.py.