Functional workflow for GitLab CI
Since GitLab doesn't have bot user to commit on repositories like GitHub
does, the user has to specify it's credentials, which will be used to
generate the report. It will also be necessary to create and add an
access token for that user so he is allowed to access the repository.