Fix HoverTool issues (esp. on boxplots)
Because the HoverTool was active on all glyphs, it would appear multiple
times when hovering a boxplot (on sometimes on a dotplot with confidence
intervals displayed). The HoverTool is now restricted to one type of
glyph per plot, and is now displayed when the mouse horizontal coords
only are matching.