Global install of vfc_ci
Python imports are now made from the Verificarlo module, and including
sigdigts (now a submodule of this repo). Other parts of the code have also
been updated accordingly (especially the workflow templates).

To set up vfc_ci over an existing install, follow the quick install
guide in the readme. The workflow of this repo has been updated with
an hardcoded install procedure.