Correct quick install process & new infos in run metadata
Using the source ... to export variables was not convenient because the
make command will probably require us to be super user. Instead
everything is copied from the bash script that can be called via make
install if we also want to DL Git submodules.

New infos have been added to the runs metadata to prepare the
generalization of the report to the multi-repositories generalization.