Support for multiple repositories in report
Now that the repository metadata is included in the run files,
integrating (multiple) Git repositories in one report is done
automatically, without having to use a CLI argument. The data is
filtered by repo by a selector on top of the webpage (this is done in
main.py in the code), and the rest works exactly the same, but using
pre-filtered data.