Lay basis for toleranche thresholds on probes values
This commit defines a format of config file, a CLI argument to link such
a file to the report, and incorporate its content into CompareRuns. This
will allow to show a visual waring when a standard deviation goes too
high or significant digits are too low compared to the desired value.