New assert system on vfc_ci
Accuracy thresholds can be optionally set with vfc_probe_assert which
works just like vfc_probe with an additional argument. When calling
tehe default vfc_probe, the accuracy threshold is set to 0, which means
it must me ignored. Finally, this value is also exported in the .csv
file when calling vfc_dump_probes.