Execution and data processing of deterministic tests
If a backend doesn't have a number of repetitions specified, it will be
assumed that it is determi. In that case, it will be executed only once,
and if it has an assert linked to it, it will be validated by a
reference run with the IEEE backend.

All data from deterministic backend can't have the same data processing
as the other backends. For this reason, data from such backends is
stored in a separated dataframe that doesn't contain all the useless
columns.

Because of that, the main "data" dataframe can now be empty (in case
we're only interested in deterministic backends), so the server has been
modified to support this case.